+++ tests/bootstrap.php
@@ -4,7 +4,7 @@ require_once __DIR__ . '/../vendor/autoload.php';
 
 define('OMEKACLI_PATH', __DIR__ . '/..');
 define('BACKUPS_DIR', getenv('HOME') . '/.omeka-cli/backups');
-define('OMEKACLI_VERSION', '0.18.0');
+define('OMEKACLI_VERSION', '1.0.0-alpha.1');
 
 use OmekaCli\Application;