+++ CHANGELOG.md
@@ -6,13 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ###Changed
-- installation infos are given as options to the `install` commands.
+- installation infos are given as options to the `install` command.
 - --no-prompt can be replaced by -n option.
 - `install` ask for missing options
 
 ### Fixed
 - Uncaught option parsing exceptions.
 - `install` default options.
+- administator_email empty with `install` command.
 
 ## [0.18.0] - 2017-08-24
 ### Added