+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Uncaught option parsing exceptions.
 - `install` default options.
 - administator_email empty with `install` command.
+- Typo in `install` command.
 
 ## [0.18.0] - 2017-08-24
 ### Added